requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-WTF
email-validator
torch
torchvision
Pillow